Refactor the FetchProperties function inside the Request File
