+++ app/properties/page.jsx
@@ -1,16 +1,5 @@
 import PropertyCard from '@/components/PropertyCard';
-
-async function fetchProperties() {
-  try {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_API_DOMAIN}/properties`);
-    if (!res.ok) {
-      throw new Error('failed to fetch properties');
-    }
-    return res.json();
-  } catch (error) {
-    console.log(error);
-  }
-}
+import { fetchProperties } from '@/utils/requests';
 
 const PropertiesPage = async () => {
   const properties = await fetchProperties();
