+++ components/HomeProperties.jsx
@@ -1,18 +1,7 @@
 // import properties from '@/properties.json';
 import PropertyCard from '@/components/PropertyCard';
 import Link from 'next/link';
-
-async function fetchProperties() {
-  try {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_API_DOMAIN}/properties`);
-    if (!res.ok) {
-      throw new Error('failed to fetch properties');
-    }
-    return res.json();
-  } catch (error) {
-    console.log(error);
-  }
-}
+import { fetchProperties } from '@/utils/requests';
 
 const HomeProperties = async () => {
   const properties = await fetchProperties();
